refactor(cliproxy): extract provider parsing and prompts into setup/providers.ts
Move ProviderId, parseProviders, PROVIDER_DEFAULTS, promptForProviders, and
promptForModel out of setup.ts into a focused submodule with colocated tests.
Adds a __proto__ prototype-chain safety regression test (matches PR #303 lesson).

setup.ts shrinks from 1619 to 1492 LOC. No exported symbol signature changes.
673 tests pass (baseline 672 + 1 new safety regression test).
